+++ app/assets/stylesheets/components/modal.scss
@@ -59,8 +59,7 @@
   top: 15px;
   right: 35px;
   color: #f1f1f1;
-  font-size: 40px;
-  font-weight: bold;
+  font-size: 30px;
   transition: 0.3s;
 }
