+++ app/javascript/packs/application.js
@@ -12,6 +12,7 @@ import { initMinuteCreation } from '../components/init_minute_creation.js';
 import { initSpeechApi } from "../components/init_speech_api.js";
 import { flatPickr } from "../plugins/flatpickr"
 import { toggleDone } from "../components/toggle_done"
+import { modal } from "../components/modal"
 
 document.addEventListener('turbolinks:load', () => {
   initJourneyChange();
@@ -19,26 +20,7 @@ document.addEventListener('turbolinks:load', () => {
   initMinuteCreation();
   flatPickr();
   toggleDone();
+  modal();
 });
 
-// Get the modal
-const modal = document.getElementById("myModal");
 
-// Get the image and insert it inside the modal - use its "alt" text as a caption
-const img = document.querySelectorAll("myImg");
-const modalImg = document.getElementById("img01");
-img.addEventListener("click", (event) => {
-
-}
-// img.onclick = function(){
-//   modal.style.display = "block";
-//   modalImg.src = this.src;
-// }
-
-// Get the <span> element that closes the modal
-const span = document.getElementsByClassName("close")[0];
-
-// When the user clicks on <span> (x), close the modal
-span.onclick = function() {
-  modal.style.display = "none";
-}
