+++ app/views/visits/_last_minute.html.erb
@@ -20,13 +20,7 @@
   </div>
 <% end %>
 
-
-<!-- The Modal -->
 <div id="myModal" class="modal">
-
-<!-- The Close Button -->
 <span class="close">&times;</span>
-
-<!-- Modal Content (The Image) -->
 <img class="modal-content" id="img01">
 </div>
